+++ src/data/import_affymetrix.py
@@ -9,6 +9,7 @@ Created on Wed Jun  9 15:44:58 2021
 import click
 import logging
 
+from src.features.illumina import IlluSNP
 from src.features.smarterdb import global_connection, SupportedChip, Location
 from src.features.affymetrix import read_Manifest
 from src.data.common import get_variant_species, update_variant, new_variant
@@ -55,6 +56,9 @@ def main(species, manifest, chip_name, version):
             alleles = sorted([record.ref_allele, record.alt_allele])
             alleles = "/".join(alleles)
 
+        # get the illumina coded snp relying on sequence
+        illusnp = IlluSNP(record.flank).toTop()
+
         # create a location object
         location = Location(
             version=version,
@@ -63,6 +67,8 @@ def main(species, manifest, chip_name, version):
             affymetrix_ab=affymetrix_ab,
             alleles=alleles,
             strand=record.strand,
+            illumina=illusnp.snp,
+            illumina_strand=illusnp.strand,
             imported_from="affymetrix",
             date=record.date,
         )